+++ crates/plurimus_widgets/src/checkbox.rs
@@ -14,6 +14,7 @@ use bevy_input_focus::tab_navigation::TabIndex;
 use plurimus_core::ratatui_core::text::Line;
 use ratatui_widgets::paragraph::Paragraph;
 
+use crate::activate::ActivateKeys;
 use plurimus_core::UiWidget;
 use plurimus_ui::UiLabel;
 use plurimus_ui::{Hovered, UiTheme};
@@ -23,7 +24,7 @@ use plurimus_ui::{LabeledQuery, StylistCache, decorate, restyle};
 /// [`checkbox_self_update`](crate::checkbox_self_update) for uncontrolled
 /// behavior.
 #[derive(Component, Debug, Clone, Copy)]
-#[require(Hovered, StylistCache)]
+#[require(Hovered, StylistCache, ActivateKeys)]
 pub struct Checkbox;
 
 /// Spawn bundle for a standard checkbox.